requirements/api.txt
# api.txt — API 전용 (표현 계층)
-r ./common.txt

# Django REST / API
djangorestframework
django-filter
django-cors-headers
django-timezone-field

# Auth / Security
djangorestframework_simplejwt
PyJWT

# API Docs
drf-yasg
uritemplate
inflection

# Server
gunicorn
gevent>=23.0.0

reportlab

# Solapi (support.messaging views)
solapi>=5.0.0
